+++ examples/expo-example/src/App.tsx
@@ -30,7 +30,7 @@ import {
   type PendingWalletSelection,
   type WalletAccount,
   type WalletActivationResult,
-} from '@0xsequence/oms-react-native-sdk';
+} from '@polygonlabs/oms-wallet-react-native';
 
 WebBrowser.maybeCompleteAuthSession();